chore: add basic integration tests
